requirements.txt
discord.py[voice]
pytz
beautifulsoup4